Вывод законов на страницу по категориям не правильный. поменял user.setStatus(statusRepo.findById((long)1).get()); на user.setStatus(statusRepo.findById(1L).get()); вход в систему теперьь работает

+++ src/main/java/com/db/auth/repos/StatusRepo.java
@@ -3,5 +3,5 @@ package com.db.auth.repos;
 import com.db.auth.model.Status;
 import org.springframework.data.repository.CrudRepository;
 
-public interface StatusRepo extends CrudRepository<Status, Integer> {
+public interface StatusRepo extends CrudRepository<Status, Long> {
 }
